+++ lib/client/client.dart
@@ -1,4 +1,5 @@
 import "dart:async";
+import 'package:fishstick_dart/commands/fortnite/account/be_eac.dart';
 import "package:logging/logging.dart";
 
 import "package:nyxx/nyxx.dart";
@@ -31,6 +32,7 @@ import "../commands/fortnite/account/login/login.dart";
 import "../commands/fortnite/account/logout.dart";
 import "../commands/fortnite/account/account.dart";
 import "../commands/fortnite/account/launch.dart";
+export "../commands/fortnite/account/be_eac.dart";
 import "../commands/fortnite/vbucks/vbucks.dart";
 import "../commands/fortnite/sac/sac.dart";
 import "../commands/fortnite/mfa/mfa.dart";
@@ -135,6 +137,7 @@ class Client {
     _commands.addCommand(logoutCommand);
     _commands.addCommand(accountCommand);
     _commands.addCommand(gameLaunchCommand);
+    _commands.addCommand(beeacCommand);
     _commands.addCommand(vbucksCommand);
     _commands.addCommand(affiliateCommand);
     _commands.addCommand(mfaCommand);